Разнёс классы по разным файлам
